backend/requirements.txt
-i https://pypi.python.org/simple
alembic==1.0.10
bcrypt==3.1.4
certifi==2018.10.15
cffi==1.11.5; python_version != '3.0.*'
chardet==3.0.4
click==7.0; python_version != '3.1.*'
flask-cors==3.0.6
flask-migrate==2.1.1
flask-script==2.0.6
flask-sqlalchemy==2.4.0
flask==1.0.2
gunicorn==19.7.1
idna==2.6
itsdangerous==1.1.0; python_version != '3.1.*'
jinja2==2.10
mako==1.0.7
markupsafe==1.1.0; python_version != '3.1.*'
pg8000==1.12.3
psycopg2-binary==2.7.5
pycparser==2.19; python_version != '3.0.*'
pyflakes==1.6.0
pyjwt==1.7.1
python-dateutil==2.7.5
python-editor==1.0.3
requests==2.22.0
six==1.11.0
sqlalchemy-utils==0.33.11
sqlalchemy==1.3.3
testing.common.database==2.0.3
testing.postgresql==1.3.0
urllib3==1.25.2
werkzeug==0.15.4
